requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.1

# Document parsing
pdfminer.six>=20221105
python-docx>=1.0.0
PyMuPDF>=1.22.0

# YAML for role presets
PyYAML>=6.0.1

# Optional OCR for scanned PDFs (requires Tesseract installed on OS)
# pdf2image>=1.17.0
# pytesseract>=0.3.10
